requirements.txt
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=-
#

accelerate~= 0.17.1
PySide6~=6.4.2
numpy~=1.24.2
torch
Pillow~=9.4.0
tqdm~=4.64.1
diffusers~=0.15.0.dev0
torchvision~=0.15.0.dev20230311+cu117
qimage2ndarray~=1.10.0
transformers~=4.26.1
opencv-python~=4.7.0.72
opencv-python~=4.7.0.72
einops~=0.6.0
compel~=0.1.10
scipy~=1.10.1
qt_material
opencv-python
controlnet_aux